Updated NMEA and ublox to get more info about the GNSS fix quality

+++ Hardware/RazorAHRS.h
@@ -352,7 +352,7 @@ inline int ConnectRazorAHRS(RAZORAHRS* pRazorAHRS, char* szCfgFilePath)
 
 	if (pRazorAHRS->bROSMode)
 	{
-		if (WriteAllRS232Port(&pRazorAHRS->RS232Port, (uint8*)outputmodebuf, strlen(outputmodebuf)) != EXIT_SUCCESS)
+		if (WriteAllRS232Port(&pRazorAHRS->RS232Port, (uint8*)outputmodebuf, (int)strlen(outputmodebuf)) != EXIT_SUCCESS)
 		{
 			printf("Unable to connect to a RazorAHRS : Initialization failure.\n");
 			CloseRS232Port(&pRazorAHRS->RS232Port);
@@ -363,7 +363,7 @@ inline int ConnectRazorAHRS(RAZORAHRS* pRazorAHRS, char* szCfgFilePath)
 			fwrite(outputmodebuf, strlen(outputmodebuf), 1, pRazorAHRS->pfSaveFile);
 			fflush(pRazorAHRS->pfSaveFile);
 		}
-		if (WriteAllRS232Port(&pRazorAHRS->RS232Port, (uint8*)streammodebuf, strlen(streammodebuf)) != EXIT_SUCCESS)
+		if (WriteAllRS232Port(&pRazorAHRS->RS232Port, (uint8*)streammodebuf, (int)strlen(streammodebuf)) != EXIT_SUCCESS)
 		{
 			printf("Unable to connect to a RazorAHRS : Initialization failure.\n");
 			CloseRS232Port(&pRazorAHRS->RS232Port);
